network-infra-health: update network infra health package
The performance of the interface count option at scale is still causing
the dashboard to be slow to load in cases where the CV cluster has a
large number of interfaces. Removing this option was not sufficient,
the option needs to be explicitly set to false.

Also addressing the content feedback from Julie in this change.

Change-Id: Iad2375956fc6923fe61421dde46362444c537a9a