network-infra-health: update network infra health package
Remove any event types from the dashboard which are not part of on-prem
CV as during configuration mode, the name of the event shows up in the
event type selector as the raw event type and not the human readable
name as the type is not present in the response from event type details.

Also remove the tags input default value of device:* which was requiring
the expansion of the set of all devices and passing these as filters to
the events service where this can cause an error at large device scale.
Instead use a default empty value which defaults to no filters which
shows all events even at scale.

Change-Id: Id216ef87fdac096d75c76d00497a3f7c80223749